Add basic UI metatype system
For object types we'll need some meta-information to implement certain
script methods like IsObjectType.

The current implementation of this system is a UI type traits system;
object types will specialize UI traits and provide static fields or
functions to opt-in to certain behaviours.

Certain aspects of the traits will be type-erased and placed into a
generic ui_type_info container at runtime, with the idea being that
objects will indicate what type they were instantiated as.

The type-info class provides a few members for querying the name of
the instantiated type and testing aspects of its inheritance hierarchy.