Initial commit of UI service and object types
This _very partially_ implements the core skeleton of the UI service
and some of the object types, providing basic functional implementation
of a few global script functions.

Note that this is still very far from complete; a few key problems
remain with regards to object allocation, management of strings, and
persistent script references.