Add size/alignment members to metatype data
Currently there's an issue whereby our destroy_object function doesn't
pass through the correct size and alignment of the type that an object
was constructed as - instead it bases that off the type that the
destroy_object function has been passed.

As we're using an object pool for these resources the size information
is crucial to find the correct bucket from which the pointer is to be
freed, so we need a way to know at runtime what the original size and
alignment requirements were of the instantiated type.

Since we keep a reference to some meta information about the type at
runtime, we can stash such data into the ui_type_info class and retrieve
it just prior to destroying the object, then feed that information
through to the deallocator.