Homogenize and improve object allocations
The experimental placement new syntax for allocating objects via a
user-supplied ui_service parameter has been removed, in favour of the
free 'new_object' and 'delete_object' functions.

The main reasoning for this is just to have one "correct" way of doing
things, rather than doing a mix-and-match all around the place.

The 'new_object' and 'delete_object' functions have been slightly
updated to now use two new public members on the ui_service class for
allocating and deallocating objects sized and aligned according to a
ui_type_info instance. In particular, 'delete_object' will now trigger
deallocation of objects according to the runtime-instantiated type of an
object rather than the passed-as type of an object.