Make delete_object non-templated
As we now rely on the runtime type of the object we don't need to know
the type of the passed-as object, and can instead just work with any
generic object pointer.

Additionally delete_object is now a no-op if a null object pointer is
supplied, to match typical destruction semantics.