Destroy objects on ui_service destruction
Previously destroying the ui_service would only trigger a deallocation
of the underlying memory for all objects without running their
destructors. This was an intentional choice as - at the time - there
weren't potentially going to be any resources managed by objects that
would need cleaning up; the idea being that hopefully any dynamic memory
allocations required by objects could be routed through the ui_service
and all quickly and efficiently cleaned up.

However while that would be ideal, it poses a few challenges - each
object is going to need a persistent script reference tied to the
lifetime of the object, which needs cleaning up. These could all be
placed in a single master table that's owned by the ui_service, but that
complicates matters in that to access the reference for an object you'd
first need to look up this other table and then index into that.

For now - or potentially forever more - we'll instead just destroy all
owned objects normally. Given that destruction of a ui_service will only
ever occur once, it's probably not all that important to be *too*
performant and this does significantly reduce the complexity required
for implementing some things.