Rework doctest integrations
Rather than placing the tests in-line with the sources and linking to a
shared library from both the test runner and the main library, we'll
simplify things by moving the tests out to the runner binary.

To do this, the runner will now link against the objects build as part
of the hyjal library target and import its link libraries and (for now)
include directories.

This removes the need for the test_registry library target to exist as
the hyjal target no longer contains tests, as such the test_runner
target has just been renamed to test.