Constrain basic_state template
Make use of the new pointer_to concept to ensure that the supplied
pointer-like type has a valid pointer_traits specialization and supports
the 'std::to_address(p)' machinery.

To cut down on the amount of repetition of repeating the template
constraints for each member function, the definitions of those members
are now inlined directly into the basic_state class definition.